- add compress texture support for Assets package - seperate loaders + parsers - changed Transcorder to use a promise instead of a runner for initialisation event
