+++ packages/basis/src/index.ts
@@ -3,5 +3,7 @@ import { BasisLoader } from './BasisLoader';
 
 export * from './Basis';
 export * from './BasisLoader';
+export * from './BasisParser';
+export * from './TranscoderWorker';
 
 extensions.add(BasisLoader);
