+++ packages/compressed-textures/src/index.ts
@@ -1,3 +1,4 @@
 export * from './const';
 export * from './resources';
 export * from './loaders';
+export * from './parsers';
